some hacker deterants

+++ zp-core/zp-extensions/galleryArticles/combiNews.php
@@ -3,7 +3,6 @@
 /**
  * @package plugins/galleryArticles
  */
-
 class Combi extends CMS {
 
 	/**
@@ -198,10 +197,10 @@ class Combi extends CMS {
 					if ($article->checkAccess()) {
 						$counter++;
 						$latestnews[$counter] = array(
-										"albumname"	 => $article->getTitle(),
-										"titlelink"	 => $article->getTitlelink(),
-										"date"			 => $article->getDateTime(),
-										"type"			 => "news",
+								"albumname" => $article->getTitle(),
+								"titlelink" => $article->getTitlelink(),
+								"date" => $article->getDateTime(),
+								"type" => "news",
 						);
 					}
 				}
@@ -218,10 +217,10 @@ class Combi extends CMS {
 						$albumdate = strftime('%Y-%m-%d %H:%M:%S', $timestamp);
 					}
 					$latestalbums[$counter] = array(
-									"albumname"	 => $tempalbum->getFileName(),
-									"titlelink"	 => $tempalbum->getTitle(),
-									"date"			 => $albumdate,
-									"type"			 => 'albums',
+							"albumname" => $tempalbum->getFileName(),
+							"titlelink" => $tempalbum->getTitle(),
+							"date" => $albumdate,
+							"type" => 'albums',
 					);
 				}
 				//$latestalbums = array_merge($latestalbums, $item);
@@ -256,42 +255,45 @@ class Combi extends CMS {
 }
 
 global $plugin_is_filter;
-enableExtension('galleryArticles', $plugin_is_filter);
 
-$obj = new Combi();
-$combi = $obj->getOldCombiNews();
-$cat = newCategory('combinews', true);
-$cat->setTitle(gettext('combiNews'));
-$cat->setDesc(gettext('Auto category for ported combi-news articles.'));
-$cat->save();
+if (class_exists('galleryArticles')) {
+	enableExtension('galleryArticles', $plugin_is_filter);
 
-foreach ($combi as $article) {
-	switch ($article['type']) {
-		case 'images':
-			$obj = newImage(array('folder' => $article['albumname'], 'filename' => $article['titlelink']), false);
-			break;
-		case 'albums':
-			$obj = newAlbum($article['albumname'], false);
-			break;
-		default:
-			$obj = NULL;
-			break;
+	$obj = new Combi();
+	$combi = $obj->getOldCombiNews();
+	$cat = newCategory('combinews', true);
+	$cat->setTitle(gettext('combiNews'));
+	$cat->setDesc(gettext('Auto category for ported combi-news articles.'));
+	$cat->save();
+
+	foreach ($combi as $article) {
+		switch ($article['type']) {
+			case 'images':
+				$obj = newImage(array('folder' => $article['albumname'], 'filename' => $article['titlelink']), false);
+				break;
+			case 'albums':
+				$obj = newAlbum($article['albumname'], false);
+				break;
+			default:
+				$obj = NULL;
+				break;
+		}
+		if ($obj && $obj->exists) {
+			$obj->setPublishDate($article['date']);
+			galleryArticles::publishArticle($obj, 'combinews');
+		}
 	}
-	if ($obj && $obj->exists) {
-		$obj->setPublishDate($article['date']);
-		galleryArticles::publishArticle($obj, 'combinews');
-	}
+	purgeOption('zenpage_combinews');
+	purgeOption('combinews-customtitle');
+	purgeOption('combinews-customtitle-imagetitles');
+	purgeOption("zenpage_combinews_sortorder");
+	purgeOption('zenpage_combinews_imagesize');
+	purgeOption('combinews-thumbnail-width');
+	purgeOption('combinews-thumbnail-height');
+	purgeOption('combinews-thumbnail-cropwidth');
+	purgeOption('combinews-thumbnail-cropheight');
+	purgeOption('combinews-thumbnail-cropx');
+	purgeOption('combinews-thumbnail-cropy');
+	purgeOption('zenpage_combinews_mode');
 }
-purgeOption('zenpage_combinews');
-purgeOption('combinews-customtitle');
-purgeOption('combinews-customtitle-imagetitles');
-purgeOption("zenpage_combinews_sortorder");
-purgeOption('zenpage_combinews_imagesize');
-purgeOption('combinews-thumbnail-width');
-purgeOption('combinews-thumbnail-height');
-purgeOption('combinews-thumbnail-cropwidth');
-purgeOption('combinews-thumbnail-cropheight');
-purgeOption('combinews-thumbnail-cropx');
-purgeOption('combinews-thumbnail-cropy');
-purgeOption('zenpage_combinews_mode');
 ?>
